Reorganizar importaciones de InputError y TextLink para utilizar la ruta compartida en las páginas de autenticación
